refactor: Make the lib position independent
This will make it easier to move this code into an `op`
