+++ src/lib.rs
@@ -26,15 +26,15 @@ mod registry;
 mod util;
 mod version;
 
-pub use crate::crate_spec::CrateSpec;
-pub use crate::dependency::Dependency;
-pub use crate::errors::*;
-pub use crate::fetch::{
+pub use crate_spec::CrateSpec;
+pub use dependency::Dependency;
+pub use errors::*;
+pub use fetch::{
     get_features_from_registry, get_latest_dependency, get_manifest_from_path,
     get_manifest_from_url, update_registry_index,
 };
-pub use crate::manifest::{find, LocalManifest, Manifest};
-pub use crate::metadata::{manifest_from_pkgid, workspace_members};
-pub use crate::registry::registry_url;
-pub use crate::util::{colorize_stderr, ColorChoice};
-pub use crate::version::{upgrade_requirement, VersionExt};
+pub use manifest::{find, LocalManifest, Manifest};
+pub use metadata::{manifest_from_pkgid, workspace_members};
+pub use registry::registry_url;
+pub use util::{colorize_stderr, ColorChoice};
+pub use version::{upgrade_requirement, VersionExt};
